% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{coerceClass}
\alias{coerceClass}
\title{Coerce select columns to a certain class}
\usage{
coerceClass(data, columns, fun = "as.factor")
}
\arguments{
\item{data}{A data.table}

\item{columns}{A character vector of columns to coerce}

\item{fun}{A function to coerce columns. Default is as.factor}
}
\value{
a data.table with columns specified in \code{columns} coerced to a
 certain class
}
\description{
Coerce select columns to a certain class
}
