% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{formatXgbProbs}
\alias{formatXgbProbs}
\title{Format prediction probabilities produced by xgboost model}
\usage{
formatXgbProbs(xgbMod, newdat, classes = LETTERS[1:4])
}
\arguments{
\item{xgbMod}{A model produced using xgboost package}

\item{newdat}{New data of class "xgb.DMatrix"}

\item{classes}{A vector of class labels}
}
\value{
a matrix of probabilities with dimensions \code{[nrow(newdat), length(classes)]}
}
\description{
Format prediction probabilities produced by xgboost model
}
