% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{multiCodeVars}
\alias{multiCodeVars}
\title{Code a factor or character vector as integer}
\usage{
multiCodeVars(fac)
}
\arguments{
\item{fac}{A factor, typically "non-numeric" levels.}
}
\value{
A converted integer vector. Values start at 0.
}
\description{
Code a factor or character vector as integer
}
