% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{processTestData}
\alias{processTestData}
\title{Format test data. Part of workflow with processTrainData}
\usage{
processTestData(dat, sc_means, sc_sds, scale_cols)
}
\arguments{
\item{dat}{A data.table containing test data}

\item{sc_means}{A vector of means produced by processTrainData}

\item{sc_sds}{A vector of standard deviations produced by processTrainData}

\item{scale_cols}{A vector of columns to scale produced by processTrainData}
}
\value{
a data.table containing the processed data to make predictions with
}
\description{
Format test data. Part of workflow with processTrainData
}
