% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{processTrainData}
\alias{processTrainData}
\title{Process training data}
\usage{
processTrainData(data, y)
}
\arguments{
\item{data}{A data.table of independent variables}

\item{y}{A data.table with one columns, your dependent variable}
}
\value{
A list containing the following elements
 \itemize{
 \item \strong{x_train} a processed training data.table
 \item \strong{y_train} a data.table with response variable matching x_train index
 \item \strong{scale_means} a vector of means used to scale training data
 \item \strong{scale_sds} a vector of standard deviations used to scale training data
 \item \strong{scale_cols} a character vector of columns that were scaled
 }
}
\description{
Process training data
}
