% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{readAndSplit}
\alias{readAndSplit}
\title{Read in training or test data}
\usage{
readAndSplit(path, target_name, name = "train")
}
\arguments{
\item{path}{A string representing a file path.}

\item{target_name}{Name of response column.}

\item{name}{Either train or test to denote type of data.}
}
\value{
Assigns two objects to the parent frame. An data.table of independent
  variables and a data.table with dependent variable.
}
\description{
Read in training or test data
}
